+++ src/main/java/io/github/offbeat_stuff/zombie_apocalypse/config/ConfigHandler.java
@@ -18,7 +18,7 @@ public class ConfigHandler {
 
   public static void load(Config config) {
     zombiesBurnInSunlight = config.zombiesBurnInSunlight;
-    ScreamHandler.setDoScream(config.doScream);
+    ScreamHandler.load(config.Scream);
 
     SpawnHandler.load(config.Spawning);
     EquipmentHandler.load(config.Equipment);
@@ -27,10 +27,22 @@ public class ConfigHandler {
   }
 
   public static void correct(Config config) {
+    correct(config.Scream);
     correct(config.Spawning);
     correct(config.Equipment);
     correct(config.ArmorTrims);
     correct(config.statusEffects);
+  }
+
+  private static void correct(ScreamConfig conf) {
+    conf.message = conf.message == null || conf.message.equals("")
+                       ? "Zombies are coming"
+                       : conf.message;
+    conf.sound = identifier(conf.sound);
+    conf.sound = VersionDependent.isSound(conf.sound) ? conf.sound
+                                                      : "entity.zombie.ambient";
+    conf.volume = Math.max(0, conf.volume);
+    conf.pitch = Math.max(0, conf.pitch);
   }
 
   private static void correct(SpawnConfig conf) {
@@ -68,7 +80,8 @@ public class ConfigHandler {
 
     conf.weaponChance = chance(conf.weaponChance);
 
-    correct(conf.enchantmentLevel);
+    correct(conf.Enchantment.level);
+    conf.Enchantment.chance = conf.Enchantment.chance;
   }
 
   private static void correct(TrimConfig conf) {